insert content int doc
