+++ client/src/components/Chat/WriterView.tsx
@@ -1,17 +1,12 @@
 import { memo, useEffect, useLayoutEffect, useMemo, useState } from 'react';
-import { useRecoilValue } from 'recoil';
-import { useParams, useLocation } from 'react-router-dom';
-import { useGetMessagesByConvoId } from 'librechat-data-provider/react-query';
-import { ChatContext, useFileMapContext } from '~/Providers';
-import MessagesView from './Messages/MessagesView';
-import { useChatHelpers, useSSE } from '~/hooks';
-import { Spinner } from '~/components/svg';
+import { useParams } from 'react-router-dom';
+
+import { ChatContext } from '~/Providers';
+
+import { useChatHelpers } from '~/hooks';
 import ChatForm from './Input/ChatForm';
-import { buildTree } from '~/utils';
-import Landing from './Landing';
-import Header from './Header';
+
 import Footer from './Footer';
-import store from '~/store';
 import WriterPresentation from './WriterPresentation';
 import { useUser } from '@clerk/clerk-react';
 import { TiptapCollabProvider } from '@hocuspocus/provider';
@@ -19,18 +14,6 @@ import * as Y from 'yjs';
 
 function WriterView({ index = 0 }: { index?: number }) {
   const { conversationId } = useParams();
-  const submissionAtIndex = useRecoilValue(store.submissionByIndex(0));
-  useSSE(submissionAtIndex);
-
-  const fileMap = useFileMapContext();
-
-  const { data: messagesTree = null, isLoading } = useGetMessagesByConvoId(conversationId ?? '', {
-    select: (data) => {
-      const dataTree = buildTree({ messages: data, fileMap });
-      return dataTree?.length === 0 ? null : dataTree ?? null;
-    },
-    enabled: !!fileMap,
-  });
 
   const chatHelpers = useChatHelpers(index, conversationId);
 
@@ -99,15 +82,6 @@ function WriterView({ index = 0 }: { index?: number }) {
         ydoc={ydoc}
         provider={provider}
       >
-        {isLoading && conversationId !== 'new' ? (
-          <div className="flex h-screen items-center justify-center">
-            <Spinner className="opacity-0" />
-          </div>
-        ) : messagesTree && messagesTree.length !== 0 ? (
-          <MessagesView messagesTree={messagesTree} Header={<Header />} />
-        ) : (
-          <Landing Header={<Header />} />
-        )}
         <div className="w-full border-t-0 pl-0 pt-2 dark:border-white/20 md:w-[calc(100%-.5rem)] md:border-t-0 md:border-transparent md:pl-0 md:pt-0 md:dark:border-transparent">
           <ChatForm index={index} />
           <Footer />
